+++ server/component/style/button/ButtonView.php
@@ -22,6 +22,27 @@ class ButtonView extends StyleView
     private $label;
 
     /**
+     * DB field 'confirmation_title' (empty string).
+     * If set a modal is shown. This will be the header of the confirmation modal.
+     */
+    private $confirmation_title;
+
+    /**
+     * DB field 'label_cancel' (empty string).
+     */
+    private $label_cancel;
+
+    /**
+     * DB field 'label_continue' (OK).
+     */
+    private $label_continue;
+
+    /**
+     * DB field 'label_message' ('Do you want to continue?').
+     */
+    private $label_message;
+
+    /**
      * DB field 'url' (empty string).
      * The target url of the button. If this field is not set, the button is not
      * rendered.
@@ -60,6 +81,10 @@ class ButtonView extends StyleView
         $this->type = $this->model->get_db_field("type", "primary");
         $this->id = $this->model->get_db_field("id", null);
         $this->data = $this->model->get_db_field("data", null);
+        $this->confirmation_title = $this->model->get_db_field("confirmation_title");
+        $this->label_cancel = $this->model->get_db_field("label_cancel");
+        $this->label_continue = $this->model->get_db_field("label_continue");
+        $this->label_message = $this->model->get_db_field("label_message");
     }
 
     /* Public Methods *********************************************************/
@@ -69,6 +94,13 @@ class ButtonView extends StyleView
      */
     public function output_content()
     {
+        $data_confirmation = array();
+        if($this->confirmation_title){
+            $data_confirmation['confirmation_title'] = $this->confirmation_title;
+            $data_confirmation['label_cancel'] = $this->label_cancel;
+            $data_confirmation['label_continue'] = $this->label_continue;
+            $data_confirmation['label_message'] = $this->label_message;
+        }
         if ($this->url == "" || $this->label == "") return;
         require __DIR__ . "/tpl_button.php";
     }